Start using unified counter
Simplify all kind of counters to `counter`.
Rework test using new api.
