+++ _examples/simple_counter_test.go
@@ -54,16 +54,16 @@ func TestSimpleCounter(t *testing.T) {
 	gometer.SetSeparator("\n")
 
 	// init simple counter and increment it 10 times.
-	inc := gometer.NewIncrementor("number_incrementor")
+	inc := gometer.NewCounter("number_incrementor")
 	for i := 0; i < 10; i++ {
-		inc.Inc()
+		inc.Add(1)
 	}
-	assert.Equal(t, int64(10), inc.Value())
+	assert.Equal(t, int64(10), inc.Get())
 
 	dec := gometer.NewCounter("number_decrementor")
 	dec.Set(5)
-	dec.Dec()
-	assert.Equal(t, int64(4), dec.Value())
+	dec.Add(-1)
+	assert.Equal(t, int64(4), dec.Get())
 
 	// write all metrics to file.
 	err = gometer.Write()
@@ -84,12 +84,12 @@ func TestSimpleCounter(t *testing.T) {
 	// check the corresponding names and values for metrics.
 	checkMetric(t, checkMetricParams{
 		name:        "number_incrementor",
-		value:       inc.Value(),
+		value:       inc.Get(),
 		metricsData: metricsData,
 	})
 	checkMetric(t, checkMetricParams{
 		name:        "number_decrementor",
-		value:       dec.Value(),
+		value:       dec.Get(),
 		metricsData: metricsData,
 	})
 }
